# TIG-1563 Improve metrics parsing performance and log csv parsing progress