--- a/src/python/tests/fixtures/cedar/barebones.csv
+++ b/src/python/tests/fixtures/cedar/barebones.csv
@@ -8,5 +8,5 @@ actor,operation,workers
 MyActor,MyOperation,2
 
 Operations
-timestamp, actor  , thread, operation, duration, outcome, n, ops, errors, size
-12345000000    , MyActor, 0     , MyOperation   , 100     , 0      , 1, 6  , 2     , 40
+timestamp,actor,thread,operation,duration,outcome,n,ops,errors,size
+12345000000,MyActor,0,MyOperation,100,0,1,6,2,40