--- a/src/python/tests/fixtures/cedar/error_outcome.csv
+++ b/src/python/tests/fixtures/cedar/error_outcome.csv
@@ -8,7 +8,7 @@ actor,operation,workers
 MyActor,MyOperation,2
 
 Operations
-timestamp, actor  , thread, operation, duration, outcome, n, ops, errors, size
-12345000000    , MyActor, 0     , MyOperation   , 100     , 0      , 1, 6  , 2     , 40
-12345000000    , MyActor, 0     , MyOperation   , 100     , 2      , 1, 6  , 2     , 40
-12345000000    , MyActor, 0     , MyOperation   , 100     , 0      , 1, 6  , 2     , 40
+timestamp,actor,thread,operation,duration,outcome,n,ops,errors,size
+12345000000,MyActor,0,MyOperation,100,0,1,6,2,40
+12345000000,MyActor,0,MyOperation,100,2,1,6,2,40
+12345000000,MyActor,0,MyOperation,100,0,1,6,2,40