--- a/src/python/tests/fixtures/cedar/two_op.csv
+++ b/src/python/tests/fixtures/cedar/two_op.csv
@@ -9,22 +9,22 @@ InsertRemove,Insert,5
 InsertRemove,Remove,5
 
 Operations
-timestamp     , actor       , thread, operation, duration, outcome, n, ops, errors, size
-66632088, InsertRemove, 0     , Insert   , 100     , 0      , 1, 6  , 2     , 40
-66632403, InsertRemove, 0     , Insert   , 310     , 0      , 1, 9  , 3     , 60
-66632661, InsertRemove, 0     , Insert   , 180     , 0      , 1, 8  , 6     , 20
-66632088, InsertRemove, 1     , Insert   , 280     , 1      , 1, 3  , 1     , 20
-66632316, InsertRemove, 1     , Insert   , 190     , 1      , 1, 4  , 1     , 30
-66632088, InsertRemove, 2     , Insert   , 30      , 0      , 1, 4  , 1     , 30
-66632245, InsertRemove, 2     , Insert   , 80      , 1      , 1, 8  , 7     , 10
-66632088, InsertRemove, 3     , Insert   , 110     , 0      , 1, 7  , 2     , 50
-66632088, InsertRemove, 4     , Insert   , 40      , 0      , 1, 9  , 0     , 90
-66632088, InsertRemove, 0     , Remove   , 100     , 0      , 1, 6  , 2     , 40
-66632403, InsertRemove, 0     , Remove   , 410     , 0      , 1, 9  , 3     , 60
-66632661, InsertRemove, 0     , Remove   , 80      , 0      , 1, 8  , 6     , 21
-66632088, InsertRemove, 4     , Remove   , 280     , 1      , 1, 3  , 2     , 24
-66632316, InsertRemove, 1     , Remove   , 123     , 1      , 1, 4  , 2     , 35
-66632088, InsertRemove, 3     , Remove   , 130     , 0      , 1, 4  , 1     , 36
-66632245, InsertRemove, 2     , Remove   , 139     , 1      , 1, 8  , 7     , 10
-66632088, InsertRemove, 2     , Remove   , 110     , 0      , 1, 7  , 2     , 12
-66632088, InsertRemove, 1     , Remove   , 20      , 0      , 1, 9  , 0     , 19
+timestamp,actor,thread,operation,duration,outcome,n,ops,errors,size
+66632088,InsertRemove,0,Insert,100,0,1,6,2,40
+66632403,InsertRemove,0,Insert,310,0,1,9,3,60
+66632661,InsertRemove,0,Insert,180,0,1,8,6,20
+66632088,InsertRemove,1,Insert,280,1,1,3,1,20
+66632316,InsertRemove,1,Insert,190,1,1,4,1,30
+66632088,InsertRemove,2,Insert,30,0,1,4,1,30
+66632245,InsertRemove,2,Insert,80,1,1,8,7,10
+66632088,InsertRemove,3,Insert,110,0,1,7,2,50
+66632088,InsertRemove,4,Insert,40,0,1,9,0,90
+66632088,InsertRemove,0,Remove,100,0,1,6,2,40
+66632403,InsertRemove,0,Remove,410,0,1,9,3,60
+66632661,InsertRemove,0,Remove,80,0,1,8,6,21
+66632088,InsertRemove,4,Remove,280,1,1,3,2,24
+66632316,InsertRemove,1,Remove,123,1,1,4,2,35
+66632088,InsertRemove,3,Remove,130,0,1,4,1,36
+66632245,InsertRemove,2,Remove,139,1,1,8,7,10
+66632088,InsertRemove,2,Remove,110,0,1,7,2,12
+66632088,InsertRemove,1,Remove,20,0,1,9,0,19