I threw in some changes to improve the perf. We should see roughly double the throughput of the first half of csv parsing. (double `tottime` of `method 'writerow' of '_csv.writer' objects`).

I haven't benchmarked csvsort yet but I'm less confident about making large improvements there. If this PR doesn't visibly improve the overall performance of metrics parsing, we may have to do something else entirely for the longer term.

```
// python -m cProfile -s tottime -m genny.cedar_report
// BEFORE
         17957893 function calls (17954266 primitive calls) in 11.364 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   637040    5.467    0.000    7.051    0.000 csv2.py:85(_parse_into_intermediate_csv)
   637040    2.391    0.000    2.391    0.000 {method 'writerow' of '_csv.writer' objects}
637057/637041    0.978    0.000    0.999    0.000 {built-in method builtins.next}
  6370400    0.575    0.000    0.575    0.000 {method 'isdigit' of 'str' objects}
  6370419    0.525    0.000    0.525    0.000 {method 'strip' of 'str' objects}
        1    0.332    0.332   11.096   11.096 cedar.py:173(split_into_actor_csv_files)
   637040    0.321    0.000    8.372    0.000 csv2.py:82(__next__)
   637040    0.273    0.000    0.273    0.000 csv2.py:109(<listcomp>)
   637041    0.112    0.000    0.112    0.000 csv2.py:273(default_columns)
1280275/1279786    0.100    0.000    0.100    0.000 {built-in method builtins.len}

// AFTER
         7486149 function calls (7482522 primitive calls) in 11.639 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1226057    4.472    0.000    4.472    0.000 {method 'writerow' of '_csv.writer' objects}
  1226057    3.817    0.000    4.007    0.000 csv2.py:85(_parse_into_intermediate_csv)
1226074/1226058    1.739    0.000    1.774    0.000 {built-in method builtins.next}
  1226057    0.608    0.000    6.388    0.000 csv2.py:82(__next__)
        1    0.582    0.582   11.445   11.445 cedar.py:173(split_into_actor_csv_files)
  2457609    0.190    0.000    0.190    0.000 {method 'rstrip' of 'str' objects}
       17    0.029    0.002    0.030    0.002 {method 'get_filename' of 'zipimport.zipimporter' objects}
     17/1    0.029    0.002    0.075    0.075 {method 'load_module' of 'zipimport.zipimporter' objects}
     9399    0.019    0.000    0.019    0.000 {built-in method _codecs.utf_8_decode}
       25    0.016    0.001    0.017    0.001 {built-in method _imp.create_dynamic}
     9399    0.016    0.000    0.034    0.000 codecs.py:319(decode)
      180    0.016    0.000    0.016    0.000 {built-in method marshal.loads}
```