@guoyr 2019-03-28
> If I'm reading the profiler output correctly, then split_into_actor_csv_files() is the slowest step.
Are you referring to the `cumtime` column? The cumulative time is the total time spent in a function including function calls. `tottime` is  the time spent outside of function calls and is the thing I we care about. Since `split_into_actor_csv_files` calls `writerow` and `_parse_into_intermediate_csv `, its cumulative time is going to be the largest. But its `tottime` should be very low.

> Why wouldn't we want to focus on that the most because it seems like any performance bottlenecks in _parse_into_intermediate_csv() could maybe be ignored by just using multiprocessing to handle each csv file.
multiprocessing could work. But it's not a trivial amount of work. Genny produces one large metrics file and the script splits it up into multiple files. We'd need to parse the one file in parallel to get benefits from multiprocessing